# account/api-docs.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "API Account",
        "description": "API para manipulação de contas de usuários do ecommerce.",
        "contact": {
            "email": "ecomm@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "API de teste"
        }
    ],
    "paths": {
        "/account": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Cadastro de conta",
                "description": "Essa rota será responsável por cadastrar uma nova conta",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Account"
                            },
                            "examples": {
                                "account": {
                                    "value": {
                                        "name": "Nanna",
                                        "email": "nannabessa@gmail.com",
                                        "password": "123456"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error creating user!"
                    },
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/account/{id}": {
            "get": {
                "summary": "Buscar uma conta",
                "description": "Essa rota será responsável por buscar uma conta",
                "operationId": "getAccounById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID retorno de uma conta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Error returning user!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deletar uma conta",
                "description": "Essa rota será responsável por deletar uma conta",
                "operationId": "deleteAccount",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID retorno de uma conta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Error deleting user!"
                    }
                }
            },
            "components": {
                "schemas": {
                    "Account": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}